refactor: image dto패키지 이동
